requirements.txt
streamlit=0.84.0
numpy==1.20.1
pandas==1.2.2
Pillow==8.1.0
matplotlib==3.3.4